refactor: LoginMember 레코드명 LoginInfo로 개선
