--- src/main/java/roomescape/domain/LoginMember.java
@@ -1,5 +1,0 @@
-package roomescape.domain;
-
-public record LoginMember(Long id, String name, String email, String role) {
-
-}
