+++ src/main/java/roomescape/dto/auth/LoginInfo.java
@@ -1,0 +1,5 @@
+package roomescape.dto.auth;
+
+public record LoginInfo(Long id, String name, String email, String role) {
+
+}
