+++ src/main/java/roomescape/repository/JdbcMemberDao.java
@@ -8,7 +8,7 @@ import org.springframework.jdbc.core.namedparam.MapSqlParameterSource;
 import org.springframework.jdbc.core.namedparam.SqlParameterSource;
 import org.springframework.jdbc.core.simple.SimpleJdbcInsert;
 import org.springframework.stereotype.Repository;
-import roomescape.domain.Member;
+import roomescape.domain.member.Member;
 
 import javax.sql.DataSource;
 import java.util.Optional;
