+++ src/main/java/roomescape/util/JwtTokenProvider.java
@@ -6,7 +6,7 @@ import io.jsonwebtoken.JwtException;
 import io.jsonwebtoken.Jwts;
 import io.jsonwebtoken.security.Keys;
 import org.springframework.stereotype.Component;
-import roomescape.domain.Member;
+import roomescape.domain.member.Member;
 
 import javax.crypto.SecretKey;
 import java.nio.charset.StandardCharsets;
